refactor utils
