+++ apps/app/src/server/service/file-uploader/aws.ts
@@ -14,12 +14,12 @@ import urljoin from 'url-join';
 
 import type { RespondOptions } from '~/server/interfaces/attachment';
 import type { IAttachmentDocument } from '~/server/models';
-import { ContentHeaders } from '~/server/routes/attachment/utils/headers';
 import loggerFactory from '~/utils/logger';
 
 import { configManager } from '../config-manager';
 
 import { AbstractFileUploader, type SaveFileParam } from './file-uploader';
+import { ContentHeaders } from './utils';
 
 
 const logger = loggerFactory('growi:service:fileUploaderAws');
